backend/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pydantic==2.8.2
SQLAlchemy==2.0.32
psycopg[binary]==3.2.9
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.7
rq==1.16.2
boto3==1.34.156
email-validator==2.2.0
python-multipart==0.0.9